changes with separate components
