+++ src/app/product.ts
@@ -1,0 +1,9 @@
+type conditionType = "new" | "used" | "discontinued";
+
+export class Product {
+  id: number;
+  name: string;
+  description: string;
+  price: number;
+  condition: conditionType;
+}
